Reorganize code
Move code from __init__ to new coordinator & helper modules.
Create a subclass of DataUpdateCoordinator.
Save cookies when config is unloaded & close API (& its session.)
When reconfiguring with new cookies file, unload config before saving
cookies so unload doesn't overwrite new cookies.
